+++ RobotCode/src/main/java/frc/lib/modules/leds/Color.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.lib.modules.leds;
 
 import java.io.File;
 import java.util.HashMap;
